requirements-dev.txt
-r requirements.txt
black
django-debug-toolbar
pytest
pytest-django
